% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumAssign.R, R/enumExtract.R
\name{[<-.enum}
\alias{[<-.enum}
\alias{[[<-.enum}
\alias{names<-.enum}
\alias{extract-enum}
\alias{[[.enum}
\alias{[.enum}
\title{Extract Enum Value}
\usage{
\method{[}{enum}(x, ...) <- value

\method{[[}{enum}(x, ...) <- value

\method{names}{enum}(x) <- value

\method{[[}{enum}(x, ..., exact = TRUE)

\method{[}{enum}(x, ..., drop = FALSE)
}
\arguments{
\item{x}{enum object from which to extract element(s)}

\item{...}{a specification of indices.}

\item{value}{an `R` object}

\item{exact}{Controls possible partial matching of `[[` when extracting
by a character vector. The default is no partial matching. Value `FALSE` allows
partial matching without any warning.}

\item{drop}{With `drop=TRUE` the result will be integer. By default subsetting
returns an `enum`.}
}
\value{
`x[[...]]` returns the value as `enum`.

For `x[...]` the default return type is `enum`. Duplicate values or `drop=TRUE` will
drop the return type to `integer`.
}
\description{
Access an enumeration by name or value.
}
\section{Functions}{
\itemize{
\item \code{`[`(enum) <- value}: Forbids enum value assignments

\item \code{`[[`(enum) <- value}: Forbids enum value assignments

\item \code{names(enum) <- value}: Forbids enum value assignments

\item \code{[}: Extract elements of an enum

}}
\note{
Unlike equivalent functions enumerations are not subset by index.
Every enumeration value can be accessed by string or the according integer
value. Subsetting *by position* is not possible for enumerations.

Unlike the equivalent function in base R this function does not throw a
warning when `exact` is `NA`.

`x[[...]]` does not accept vector arguments.
}
\examples{
e <- enum(a = 1, x = 2, c = 4)
e[[2]]
#> x
#> 2
e[["x"]]
#> x
#> 2
tryCatch(e[[3]], error = function(e) e ) # does not exist
#> Attempt to access non-existing enum value
}
