% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{inEnum}
\alias{inEnum}
\title{inEnum}
\usage{
inEnum(Needle, Haystack)
}
\arguments{
\item{Needle}{an integer or string representing an element of the
enumeration `Haystack`}

\item{Haystack}{an enumeration class `enum`}
}
\value{
TRUE/FALSE
}
\description{
Is `Needle` element of an enum `Haystack?
}
\examples{
myEnum <- enum(a=1, b=3, c=5)
inEnum(3, myEnum)   # TRUE
inEnum("b", myEnum) # TRUE
inEnum(2, myEnum)   # FALSE, however ...
inEnum(myEnum[2], myEnum)   # TRUE
inEnum(myEnum["c"], myEnum) # TRUE
}
