% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{is.enum}
\alias{is.enum}
\title{is.enum}
\usage{
is.enum(x)
}
\arguments{
\item{x}{any `R` object}
}
\value{
`TRUE` if object is a valid enum, otherwise `FALSE`.
}
\description{
Test if object is inherited from class `enum`. Moreover, the function checks if
the object can be coerced to `enum` and verifies if it is indeed a valid `enum`.
}
\examples{
is.enum(enum(a = 1, b = 2)) # TRUE
is.enum(c(a = 1, b = 2))    # FALSE
}
