% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plainEnum-package.R
\docType{package}
\name{plainEnum-package}
\alias{plainEnum}
\alias{plainEnum-package}
\title{plainEnum: Enumerations based on Atomic Vectors}
\description{
An enumeration type (or enum type) is a value type defined by a set of named constants of the underlying integral numeric type. This package emulates/simulates that in R.
}
\details{
An enumeration type (or enum type) is a value type defined by a set of named constants
of the underlying integral numeric type. To define an enumeration type, use
[enum()] and specify the names and values of enum members.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/}
  \item Report bugs at \url{https://github.com/.../issues}
}

}
\author{
\strong{Maintainer}: Jan Seifert \email{enum@sigurdjanson.anonaddy.com}

}
\keyword{internal}
