% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{print.enum}
\alias{print.enum}
\title{print.enum}
\usage{
\method{print}{enum}(x, ...)
}
\arguments{
\item{x}{an `enum` object to print}

\item{...}{further arguments passed to or from other methods.}
}
\value{
`[invisible](x)`
}
\description{
Prints its argument and returns it invisibly (via `[invisible](x)`).
}
\examples{
print(enum(a = 1, b = 2))
}
