% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\name{strip}
\alias{strip}
\alias{strip.default}
\title{strip}
\usage{
strip(obj)

\method{strip}{default}(obj)
}
\arguments{
\item{obj}{any object}
}
\value{
`strip()` removes all names, class, comments and any other attributes and returns
the result.
}
\description{
Strip an object from all it's attributes.
}
\section{Methods (by class)}{
\itemize{
\item \code{strip(default)}: Strip any object from all it's attributes.

}}
\examples{
strip(enum(a = 1, b = 2)) # 1:2
strip(c(a = 1, b = 2))    # 1:2
}
